+++ app/src/main/AndroidManifest.xml
@@ -1,9 +1,9 @@
-<?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.ezhil.techtoaid.ezhilprogramming">
+    package="com.ezhil.handbook">
 
     <uses-permission android:name="android.permission.INTERNET"></uses-permission>
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"></uses-permission>
+
     <application
         android:allowBackup="true"
         android:icon="@drawable/ezhilico"
@@ -37,5 +37,6 @@
         <activity android:name=".Loadexamplehtml"></activity>
 
     </application>
+
 
 </manifest>
