+++ app/src/main/java/com/ezhil/handbook/DownloadInstall.java
@@ -7,7 +7,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.*/
 
-package com.ezhil.techtoaid.ezhilprogramming;
+package com.ezhil.handbook;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -17,17 +17,19 @@ import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
+import com.ezhil.handbook.R;
 
-public class EzhilExamples_Comparison extends AppCompatActivity {
-    ListView listView;
-    String title1="oppiduthal";
-    String title2="melum_sila_oppidukal";
-    String title3="onrukku_merpatta_oppidukal";
-    String[] items={"ஒப்பிடுதல்","மேலும் சில ஒப்பீடுகள்","ஒன்றுக்கு மேற்ப்பட்ட ஒப்பீடுகள்"};
+
+public class DownloadInstall extends AppCompatActivity {
+        ListView listView;
+        String title1="installing_ezhil(3)";
+        String title2="inayathai_payanpaduthum_ezhil(3.2)";
+        String title3="python_vazhiyil(3.3)";
+        String[] items={"எழிலை நிறுவுதல் (Installing Ezhil)","இணையத்தில் எழில்","பைத்தானில்(python) எழில்"};
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.ezhilexamples_comparison);
-        listView=(ListView)findViewById(R.id.listview51);
+        setContentView(R.layout.downloadinstall);
+        listView=(ListView)findViewById(R.id.listview2);
         ArrayAdapter<String> adapter= new ArrayAdapter<String>(this,android.R.layout.simple_list_item_1,android.R.id.text1,items);
         listView.setAdapter(adapter);
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
@@ -37,24 +39,24 @@ public class EzhilExamples_Comparison extends AppCompatActivity {
                 switch (position)
                 {
                     case 0:
-                        Intent intent=new Intent(EzhilExamples_Comparison.this,Loadhtml.class);
+                        Intent intent=new Intent(DownloadInstall.this,Loadhtml.class);
                         intent.putExtra("file",title1);
                         startActivity(intent);
                         break;
                     case 1:
-                        Intent intent1=new Intent(EzhilExamples_Comparison.this,Loadhtml.class);
+                        Intent intent1=new Intent(DownloadInstall.this,Loadhtml.class);
                         intent1.putExtra("file",title2);
                         startActivity(intent1);
                         break;
                     case 2:
-                        Intent intent2=new Intent(EzhilExamples_Comparison.this,Loadhtml.class);
+                        Intent intent2=new Intent(DownloadInstall.this,Loadhtml.class);
                         intent2.putExtra("file",title3);
                         startActivity(intent2);
                         break;
                 }
-
             }
         });
 
     }
-}
+    }
+
