+++ app/src/main/java/com/ezhil/handbook/EzhilExamples_Examples.java
@@ -7,7 +7,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.*/
 
-package com.ezhil.techtoaid.ezhilprogramming;
+package com.ezhil.handbook;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -16,7 +16,8 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
-import android.widget.Toast;
+
+import com.ezhil.handbook.R;
 
 public class EzhilExamples_Examples extends AppCompatActivity {
     String[] listitems={"குறிசொற்க்கள்(keywords)",
