+++ app/src/main/java/com/ezhil/handbook/StartEzhil_Drawing.java
@@ -7,7 +7,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.*/
 
-package com.ezhil.techtoaid.ezhilprogramming;
+package com.ezhil.handbook;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -17,17 +17,19 @@ import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
+import com.ezhil.handbook.R;
 
-public class DownloadInstall extends AppCompatActivity {
-        ListView listView;
-        String title1="installing_ezhil(3)";
-        String title2="inayathai_payanpaduthum_ezhil(3.2)";
-        String title3="python_vazhiyil(3.3)";
-        String[] items={"எழிலை நிறுவுதல் (Installing Ezhil)","இணையத்தில் எழில்","பைத்தானில்(python) எழில்"};
+
+public class StartEzhil_Drawing extends AppCompatActivity {
+    ListView listView;
+    String title1="varainthu_paarpomaa";
+    String title2="nirangal&vattangal";
+    String title3="payirchi(nirangal_vattangal)";
+    String[] items={"வரைந்து பார்ப்போமா","நிரங்கள் & வட்டங்கள்","பயிற்சி"};
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.downloadinstall);
-        listView=(ListView)findViewById(R.id.listview2);
+        setContentView(R.layout.startezhil_drawing);
+        listView=(ListView)findViewById(R.id.listview43);
         ArrayAdapter<String> adapter= new ArrayAdapter<String>(this,android.R.layout.simple_list_item_1,android.R.id.text1,items);
         listView.setAdapter(adapter);
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
@@ -37,17 +39,17 @@ public class DownloadInstall extends AppCompatActivity {
                 switch (position)
                 {
                     case 0:
-                        Intent intent=new Intent(DownloadInstall.this,Loadhtml.class);
-                        intent.putExtra("file",title1);
-                        startActivity(intent);
-                        break;
+                    Intent intent=new Intent(StartEzhil_Drawing.this,Loadhtml.class);
+                    intent.putExtra("file",title1);
+                    startActivity(intent);
+                    break;
                     case 1:
-                        Intent intent1=new Intent(DownloadInstall.this,Loadhtml.class);
+                        Intent intent1=new Intent(StartEzhil_Drawing.this,Loadhtml.class);
                         intent1.putExtra("file",title2);
                         startActivity(intent1);
                         break;
                     case 2:
-                        Intent intent2=new Intent(DownloadInstall.this,Loadhtml.class);
+                        Intent intent2=new Intent(StartEzhil_Drawing.this,Loadhtml.class);
                         intent2.putExtra("file",title3);
                         startActivity(intent2);
                         break;
@@ -56,5 +58,4 @@ public class DownloadInstall extends AppCompatActivity {
         });
 
     }
-    }
-
+}
