+++ app/src/main/java/com/ezhil/handbook/StartEzhil_Strings.java
@@ -7,7 +7,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.*/
 
-package com.ezhil.techtoaid.ezhilprogramming;
+package com.ezhil.handbook;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -17,17 +17,18 @@ import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
+import com.ezhil.handbook.R;
 
-public class StartEzhil_Drawing extends AppCompatActivity {
+public class StartEzhil_Strings extends AppCompatActivity {
     ListView listView;
-    String title1="varainthu_paarpomaa";
-    String title2="nirangal&vattangal";
-    String title3="payirchi(nirangal_vattangal)";
-    String[] items={"வரைந்து பார்ப்போமா","நிரங்கள் & வட்டங்கள்","பயிற்சி"};
+    String title1="saram_neelam";
+    String title2="saram_inaikka";
+    String title3="thedal&saram_maatruthal";
+    String[] items={"சரம் நீளம் (String length)","சரம் இணைக்க(String concatenation)","தேடல் & சரங்களை மாற்றுதல்(Searching)"};
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.startezhil_drawing);
-        listView=(ListView)findViewById(R.id.listview43);
+        setContentView(R.layout.startezhil_strings);
+        listView=(ListView)findViewById(R.id.listview41);
         ArrayAdapter<String> adapter= new ArrayAdapter<String>(this,android.R.layout.simple_list_item_1,android.R.id.text1,items);
         listView.setAdapter(adapter);
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
@@ -37,17 +38,17 @@ public class StartEzhil_Drawing extends AppCompatActivity {
                 switch (position)
                 {
                     case 0:
-                    Intent intent=new Intent(StartEzhil_Drawing.this,Loadhtml.class);
-                    intent.putExtra("file",title1);
-                    startActivity(intent);
-                    break;
+                        Intent intent=new Intent(StartEzhil_Strings.this,Loadhtml.class);
+                        intent.putExtra("file",title1);
+                        startActivity(intent);
+                        break;
                     case 1:
-                        Intent intent1=new Intent(StartEzhil_Drawing.this,Loadhtml.class);
+                        Intent intent1=new Intent(StartEzhil_Strings.this,Loadhtml.class);
                         intent1.putExtra("file",title2);
                         startActivity(intent1);
                         break;
                     case 2:
-                        Intent intent2=new Intent(StartEzhil_Drawing.this,Loadhtml.class);
+                        Intent intent2=new Intent(StartEzhil_Strings.this,Loadhtml.class);
                         intent2.putExtra("file",title3);
                         startActivity(intent2);
                         break;
