--- app/src/main/java/com/ezhil/techtoaid/ezhilprogramming/Aboutweb.java
@@ -1,30 +1,0 @@
-/*Copyright 2017 Karunakaran.G Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License. You may obtain a copy of the License at
-http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.*/
-
-package com.ezhil.techtoaid.ezhilprogramming;
-
-import android.net.Uri;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.webkit.WebView;
-
-
-public class Aboutweb extends AppCompatActivity {
-    WebView wview;
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.aboutweb);
-        wview=(WebView)findViewById(R.id.aboutweb);
-        wview.setWebViewClient(null);
-        wview.getSettings().setJavaScriptEnabled(true);
-        Bundle extras=getIntent().getExtras();
-        String url=extras.getString("url");
-        wview.loadUrl(url);
-    }
-}
