+++ app/src/main/res/layout/loadexamplehtml.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Web">
+    tools:context="com.ezhil.handbook.Web">
 
     <WebView
         android:id="@+id/examplehtml"
